pull out serialization to base
