+++ param_persist/agents/sqlalchemy_agent.py
@@ -3,13 +3,10 @@ The SqlAlchemy Agent.
 
 This file was created on August 05, 2020
 """
-
-import importlib
 import json
 import logging
 import uuid
 
-from param.serializer import JSONSerialization
 from sqlalchemy.orm import sessionmaker
 
 from param_persist.agents.base import AgentBase
@@ -62,7 +59,7 @@ class SqlAlchemyAgent(AgentBase):
         db_session = kwargs.get('db_session', None)
 
         # Serialize data using param JSONSerialization class
-        serialized_param = json.loads(JSONSerialization.serialize_parameters(instance))
+        serialized_param = self.get_serialized_param(instance)
 
         # Remove name since we don't need it
         serialized_param.pop('name')
@@ -145,7 +142,7 @@ class SqlAlchemyAgent(AgentBase):
         instance_model = db_session.query(InstanceModel).get(instance_id)
         if instance_model is None:
             raise RuntimeError(f'Parameterized instance with id "{instance_id}" does not exist.')
-        serialized_param = json.loads(JSONSerialization.serialize_parameters(instance))
+        serialized_param = self.get_serialized_param(instance)
         serialized_param.pop('name')
 
         param_models_in_db = dict()
@@ -173,94 +170,3 @@ class SqlAlchemyAgent(AgentBase):
         db_session.commit()
 
         return instance_id
-
-    @staticmethod
-    def get_param_names(param_object):
-        """
-        Return list of all the names in param.
-        """
-        param_items = param_object.param.get_param_values()
-        param_names = list()
-        for item in param_items:
-            param_names.append(item[0])
-
-        return param_names
-
-    @staticmethod
-    def get_param_object_from_instance(instance_model):
-        """
-        Given an instance_model, return an associated param object.
-        """
-        # Getting param_object
-        try:
-            class_base_path, class_name = instance_model.class_path.rsplit('.', 1)
-            param_module = importlib.import_module(class_base_path)
-            parameterized_class = getattr(param_module, class_name)
-        except ImportError:
-            raise RuntimeError(f'Defined param class "class_path" was not importable.'
-                               f' Given path is "{instance_model.class_path}"')
-
-        param_object = parameterized_class()
-
-        return param_object
-
-    def update_param_object(self, param_object, serialized_data):
-        """
-        Update param_object data with given serialized data. Use JSONSerialization from param to deserialize the data.
-        """
-        # Update param object using deserialized data.
-        param_names = self.get_param_names(param_object)
-
-        # Remove extra args if needed.
-        keys_to_remove = list()
-        for key, _ in serialized_data.items():
-            if key not in param_names:
-                keys_to_remove.append(key)
-
-        for key in keys_to_remove:
-            serialized_data.pop(key)
-
-        # Deserialize param data
-        param_model_deserialize = JSONSerialization. \
-            deserialize_parameters(pobj=param_object, serialization=json.dumps(serialized_data))
-
-        for key, value in param_model_deserialize.items():
-            setattr(param_object, key, value)
-
-        return param_object
-
-    @staticmethod
-    def load_serialized_data_from_param_model(param_models):
-        """
-        Load serialized data from param models in appropriated format.
-        """
-        param_models_json = [x.value for x in param_models]
-        # Create serialized dictionary data in param serialized format to deserialize
-        param_model_serialized_data = dict()
-
-        for item in param_models_json:
-            item = json.loads(item)
-            param_model_serialized_data[item['name']] = item['value']
-
-        return param_model_serialized_data
-
-    @staticmethod
-    def get_type_from_param_instance(instance, key):
-        """
-        Get the associated type from a key of an instance from the database.
-        """
-        type_string = '.'.join([type(instance.param.__getitem__(key)).__module__,
-                                type(instance.param.__getitem__(key)).__name__]
-                               )
-        return type_string
-
-    @staticmethod
-    def get_class_path_from_param_instance(instance):
-        """
-        Get the associated class path of an instance from the database.
-        """
-        class_path = ".".join([
-            instance.__module__,
-            instance.__class__.__name__,
-        ])
-        return class_path
